% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calendar_data}
\alias{calendar_data}
\title{Get data for calendar generation}
\usage{
calendar_data()
}
\value{
tibble
}
\description{
Get data for calendar generation
}
\examples{
# ds <- calendar_data()
}
