% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{data_mod}
\alias{data_mod}
\title{Simple data modification}
\usage{
data_mod(data)
}
\arguments{
\item{data}{data set}
}
\value{
tibble
}
\description{
Simple data modification
}
\examples{
# data <- ds
}
