% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{export_ical}
\alias{export_ical}
\title{Creates}
\usage{
export_ical(
  data,
  assessor = "hvem_indtaster",
  dir = here::here("data/"),
  base.name = "laties_followup"
)
}
\arguments{
\item{data}{data set}

\item{assessor}{assessor var}

\item{dir}{output dir}

\item{base.name}{file base name}
}
\description{
Creates
}
\examples{
# data <- ds |> data_mod()
# data |> export_ical()
}
