% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{git_commit_push}
\alias{git_commit_push}
\title{Commit and push .ics calendar file}
\usage{
git_commit_push(f.path, c.message = paste("calendar update", Sys.Date()))
}
\arguments{
\item{f.path}{file paths(s)}

\item{c.message}{commit message}
}
\description{
Commit and push .ics calendar file
}
