% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{glue_if}
\alias{glue_if}
\title{Small wrapper for if_else in glue string}
\usage{
glue_if(data)
}
\arguments{
\item{data}{vector}
}
\value{
glue string
}
\description{
Small wrapper for if_else in glue string
}
