% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laties.support-package.R
\docType{package}
\name{laties.support-package}
\alias{laties.support}
\alias{laties.support-package}
\title{laties.support: Support functions for the LATIES project}
\description{
Support functions for the LATIES project.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/agdamsbo/laties.support}
  \item Report bugs at \url{https://github.com/agdamsbo/laties.support/issues}
}

}
\author{
\strong{Maintainer}: Andreas Gammelgaard Damsbo \email{agdamsbo@clin.au.dk} (\href{https://orcid.org/0000-0002-7559-1154}{ORCID})

}
\keyword{internal}
